Keccak.jl Documentation

Keccak.keccak_p — Function

keccak_p(state, ::Val{nrounds})

The Kᴇᴄᴄᴀᴋ-p permutation function as defined in FIPS-202, with the number of rounds given by nrounds. The input state as well as the returned updated state are 25-tuples of unsigned integers of type T. Their LSB-first serialization corresponds to the bitstrings S and S′ in FIPS-202; its length b is determined by the type T. E.g. for T==UInt64, one obtains the commonly used b = 25 ⋅ 64 = 1600.

For example, for b=800 (hence T==UInt32), S=1000...000 corresponds to state == (0x00000001, (0x00000000 for _ in 2:25)...).

Alternatively, the state can contain SIMD.Vecs of unsigned integers to perform a parallel transformation of multiple state instances. E.g, for state::NTuple{25,SIMD.Vec{4,UInt64}}, it holds that

julia> SIMD.Vec.((keccak_p(map(s -> s[n], state)) for n in 1:4)...) == keccak_p(state)
true

Internals

Helpers

Keccak.ℓ — Function

ℓ(::Type{T<:Unsigned})

Returns the base-2 logarithm of the bitwidth of values of type T.

Examples

julia> ℓ(UInt32)
5

julia> 2^5
32

Keccak.rc — Function

rc(t)

Evaluate t rounds of a linear feedback shift register and return the leading bit after the final round as specified in Algorithm 5 of FIPS-202.

This function is used for computing the round constants by round_consts.

Keccak.round_consts — Function

round_consts(::Type{T}, ::Val{nrounds}=Val(12+2ℓ(T)))

Compute the Keccak round constants for lane length w corresponding to the bitwidth of T and the given number fo rounds nrounds.

The return value is a tuple of Ts containing the round constants for round indices 12+2ℓ(T)-nrounds to 12+2ℓ(T)-1 in that order. The default for nrounds results in the first round index being 0.

Note

It is valid to have nrounds>12+2ℓ(T), so that round index starts negative.

Examples

julia> round_consts(UInt16)
(0x0001, 0x8082, 0x808a, 0x8000, 0x808b, 0x0001, 0x8081, 0x8009, 0x008a, 0x0088, 0x8009, 0x000a, 0x808b, 0x008b, 0x8089, 0x8003, 0x8002, 0x0080, 0x800a, 0x000a)

julia> round_consts(UInt8)
(0x01, 0x82, 0x8a, 0x00, 0x8b, 0x01, 0x81, 0x09, 0x8a, 0x88, 0x09, 0x0a, 0x8b, 0x8b, 0x89, 0x03, 0x02, 0x80)

julia> round_consts(UInt8, Val(10))
(0x8a, 0x88, 0x09, 0x0a, 0x8b, 0x8b, 0x89, 0x03, 0x02, 0x80)

julia> round_consts(UInt8, Val(20))
(0x02, 0x8a, 0x01, 0x82, 0x8a, 0x00, 0x8b, 0x01, 0x81, 0x09, 0x8a, 0x88, 0x09, 0x0a, 0x8b, 0x8b, 0x89, 0x03, 0x02, 0x80)

Kᴇᴄᴄᴀᴋ step mappings

These functions follow the specification in section 3.2 of FIPS-202.

Keccak.ι — Function

ι(state, RC)

Perform the step mapping ι (round-dependent modification of lane[0,0]) of the Keccak algorithm and return the updated state.

Instead of the round index as in FIPS-202, the corresponding round constant RC is expected as the second argument.

The state is a 25-tuple of unsigned integers, related to state array A in FIPS-202 by

A[x,y,z] == state[5y+x+1] >> z & 1

for x = 0,...,4, y = 0,...,4 and z = 0,...,w-1, where w is the bitwidth of the integers in state.

Alternatively, the state can contain SIMD.Vecs of unsigned integers to perform a parallel transformation of multiple state instances.

Keccak.θ — Function

θ(state)

Perform the step mapping θ of the Keccak algorithm and return the updated state.

The state is a 25-tuple of unsigned integers, related to state array A in FIPS-202 by

A[x,y,z] == state[5y+x+1] >> z & 1

for x = 0,...,4, y = 0,...,4 and z = 0,...,w-1, where w is the bitwidth of the integers in state.

Alternatively, the state can contain SIMD.Vecs of unsigned integers to perform a parallel transformation of multiple state instances.

Keccak.π — Function

π(state)

Perform the step mapping π (permutation of the lanes) of the Keccak algorithm and return the updated state.

The state is a 25-tuple of unsigned integers, related to state array A in FIPS-202 by

A[x,y,z] == state[5y+x+1] >> z & 1

for x = 0,...,4, y = 0,...,4 and z = 0,...,w-1, where w is the bitwidth of the integers in state.

Alternatively, the state can contain SIMD.Vecs of unsigned integers to perform a parallel transformation of multiple state instances.

Keccak.χ — Function

χ(state)

Perform the nonlinear step mapping χ of the Keccak algorithm and return the updated state.

The state is a 25-tuple of unsigned integers, related to state array A in FIPS-202 by

A[x,y,z] == state[5y+x+1] >> z & 1

for x = 0,...,4, y = 0,...,4 and z = 0,...,w-1, where w is the bitwidth of the integers in state.

Alternatively, the state can contain SIMD.Vecs of unsigned integers to perform a parallel transformation of multiple state instances.

Keccak.ρ — Function

ρ(state)

Perform the step mapping ρ (bit rotation in each lane) of the Keccak algorithm and return the updated state.

The state is a 25-tuple of unsigned integers, related to state array A in FIPS-202 by

A[x,y,z] == state[5y+x+1] >> z & 1

for x = 0,...,4, y = 0,...,4 and z = 0,...,w-1, where w is the bitwidth of the integers in state.

Alternatively, the state can contain SIMD.Vecs of unsigned integers to perform a parallel transformation of multiple state instances.